### 2020-06-08 #2310 @whedon
[PRE REVIEW]: A parallel global multiobjective framework for optimization: pagmo
**Submitting author:** @bluescarni (<a href="http://orcid.org/0000-0002-3491-1828">Francesco Biscani</a>)
**Repository:** <a href="https://github.com/esa/pagmo2-paper" target ="_blank">https://github.com/esa/pagmo2-paper</a>
**Version:** v2.15.0
