### 2020-06-08 #2310 @arfon
:wave: @bluescarni - thanks for your submission to JOSS. Could you please explain where the software is associated with this submission? We prefer the JOSS paper to be housed together with the software if at all possible.

Additionally, as described in [our blog post](https://blog.joss.theoj.org/2020/05/reopening-joss) announcing the reopening of JOSS, we're currently working in a "reduced service mode", limiting the number of papers assigned to any individual editor.

Since reopening JOSS earlier a week or so ago we've had > 70 papers submitted and as such, yours has been put in our backlog that we will be working through over the coming weeks and months.

Thanks in advance for your patience!
