### 2020-06-09 #2310 @bluescarni
Hi @arfon

the software lives in two repositories:

https://github.com/esa/pagmo2
https://github.com/esa/pygmo2

Would you like me to move the paper in a branch in the first repo?
