### 2020-06-09 #2310 @arfon
Got it, thanks @bluescarni. I think this is OK as it seems like we're going to want to review both packages as part of this review.

While I look for an editor, please take a look at the DOI fixes that Whedon suggested above (https://github.com/openjournals/joss-reviews/issues/2310#issuecomment-640899613).
