### 2020-06-09 #2310 @arfon
@eloisabentivegna - what do you think about taking this submission on as one of your first papers? It is _slightly_ unusual in that the author is submitting two packages (https://github.com/esa/pagmo2 & https://github.com/esa/pygmo2) as part of the same submission which look to be C++ and Python implementations of the same library. As a result, the paper is in a standalone repository which will require a little manual work by `@openjournals/dev` at the end of the review but we've handled similar things before.
