+++ paper.bib
@@ -38,6 +38,7 @@
   number={10-11},
   pages={555--571},
   year={2010},
+  doi={10.1016/j.parco.2010.04.002},
   publisher={Elsevier}
 }
 
@@ -85,7 +86,7 @@
       volume={17},
       pages={261--272},
       adsurl = {https://rdcu.be/b08Wh},
-      doi = {https://doi.org/10.1038/s41592-019-0686-2},
+      doi = {10.1038/s41592-019-0686-2},
 }
 
 @article{storn1997differential,
@@ -107,6 +108,7 @@
   number={2},
   pages={22--30},
   year={2011},
+  doi={10.1109/mcse.2011.37},
   publisher={IEEE Computer Society}
 }
 
