### 2020-06-09 #2310 @bluescarni
> While I look for an editor, please take a look at the DOI fixes that Whedon suggested above ([#2310 (comment)](https://github.com/openjournals/joss-reviews/issues/2310#issuecomment-640899613)).

Cheers, I think it should be fixed now.
