### 2020-06-09 #2310 @arfon
Excellent, thanks @eloisabentivegna. One minor thing, Whedon only responds to comments at the start of a post or as standalone comments. i.e. if you now type `@whedon assign @eloisabentivegna as editor` in its own comment then Whedon will assign you.
