### 2020-06-15 #2338 @eloisabentivegna
Dear @dgoldri25 and @jangmys, thanks for agreeing to review this submission! As you can see, I have started the review process, whereby the original issue has been closed and a new (this) one has been opened, with further directions for you. Please take a moment to go over the instructions and checklists above, and let me know if anything is unclear.

Please notice that this is a slightly unusual submission with respect to standard JOSS practice, in that you will not find the source code under the paper repository above, as the submission consists of two separate packages:

https://github.com/esa/pagmo2
 https://github.com/esa/pygmo2

This should not impact the review process, but let me know if you need further clarifications.

I look forward to your comments!
