### 2020-07-29 #2338 @jangmys
Concerning the point "**State of the field**":

The paper (which is globally very well written) starts directly with the general formulation of the optimization problem.
While the "Summary" provides some general motivation for designing (parallel) frameworks for multi-objective optimization, I think it would be useful and important to give an overview of existing frameworks and to explain what differentiates pagmo/pygmo from them : Why should I choose pagmo/pygmo instead of, say, jMetal, jMetalPy, DEAP, HeuristicLab, ParadisEO, Mallba or Opt4j (plus NLOpt) ?

I am aware that a complete comparison of functionalities/documentation/etc is a huge task which is certainly beyond the scope of this paper - however at least a summary comparison with existing frameworks would clearly enhance the quality of the paper and give interested readers additional motivation to use pagmo. Maybe this comparision could be restricted to frameworks that offer some parallel processing support... In the (now 8 years old) paper by Parejo, José Antonio, et al., "Metaheuristic optimization frameworks: a survey and benchmarking.", Soft Computing 16.3 (2012): 527-561 (https://core.ac.uk/reader/51388224) the authors perform a complete comparative study of metaheuristic optimization frameworks. The results of this study are (at least partially) outdated, but it could still be a good starting point (and/or a reference to give to the reader!).
