### 2020-07-29 #2338 @jangmys
The paper says "The parallel evaluation can be performed by multiple threads, processes, nodes in an HPC cluster or even by **GPU devices** (via, e.g., OpenCL or CUDA)" (p.3) - I could not find a "GPU batch evaluator" in the code (please help me out if I missed it). Of course, conceptually, a batch of solutions could be evaluated on a GPU - technically, however, there are some challenges (for example: (1) no std library in device code -> DS for individuals on the device?; (2) need to explicitly copy problem data to the device and separate implementation of __device__ fitness function (alternatively "Unified Memory" + __host__ __device__ lambdas...which are still experimental: performance?); (3) thread-data mapping on the device (only 1 thread/fitness eval or possibility of parallelizing fitness evaluation itself?. If a GPUbatchEval has already been implemented (and I just didn't see it), the paper should provide some explanations and instructions for using GPU acceleration should appear in the documentation... otherwise, if GPU support for batch evaluation is future work, the paper should clarify this.
Concerning a "unified" heterogeneous programming approach that fits well into pagmo's design, I think intel OneAPI/SYCL/Data Parallel C++ is an interesting alternative to CUDA/OpenCL.
