### 2020-07-29 #2338 @jangmys
Minor remarks (typos):
Summary: cuncurrently -> concurrently
p.2: non linear ; non linearly -> non-linear ; non-linearly
p.3: Cuncurrent fitness -> Concurrent
