### 2020-07-29 #2338 @jangmys
I was able to install pygmo/pagmo easily, following the instructions, and the provided tutorials helped me to get started quickly.
While I haven't used the software myself before doing this review, some colleagues in my department have been using pagmo/pygmo in their research for some time not - as far as I know, the fact that pagmo/pygmo is well documented and very accessible has been an important factor in their choice. An active community of users and the possibility to find help in case of difficulties and/or exchange with the developers (e.g. Gitter) is also a very positive point. Overall, the pagmo/pymgo framework is a valuable and (I've got the impression) already quite well established optimization tool.

Therefore, in order to increase the software's visibility and allow researchers to properly cite pagmo/pygmo (the last (only?) pagmo-related paper I found is from 2010, "A global optimization toolbox for massively parallel engineering optimisation"), publishing this paper in JOSS definitely makes sense.
However, before I can recommend accepting the article, I think it is important that the issues I mentioned above are addressed.
