### 2020-07-31 #2338 @eloisabentivegna
Thanks for your comments, @jangmys!

Could you raise any future issues in the respective repositories? This will ensure your suggestions will enter the package histories and be properly credited. You can create a mention here by using this issue's URL in a repository's issue, if you wish (see https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#guiding-principles).
