### 2020-08-10 #2338 @eloisabentivegna
@dgoldri25, do you concur with @jangmys' suggestions regarding a comparison with the state of the art? Is this why you have left the corresponding box unticked?
