### 2020-08-11 #2338 @dgoldri25
@eloisabentivegna I second @jangmys' suggestion. While I found the paper to be high quality overall, the review of the state of the art should be expanded.
