+++ paper.bib
@@ -56,6 +56,13 @@
    title = {pybind11 -- Seamless operability between C++11 and Python}
 }
 
+@misc{johnson:2014,
+  title={The NLopt nonlinear-optimization package},
+  author={Johnson, Steven G},
+  year={2014},
+  note = {http://github.com/stevengj/nlopt}
+}
+
 @misc{ipyparallel,
   author = {Min Ragan-Kelley},
   title = {ipyparallel: Interactive Parallel Computing in Python},
@@ -132,4 +139,45 @@
   author={Hagberg, Aric and Swart, Pieter and S Chult, Daniel},
   year={2008},
   institution={Los Alamos National Lab.(LANL), Los Alamos, NM (United States)}
+}
+
+@article{durillo:2011,
+  title={jMetal: A Java framework for multi-objective optimization},
+  author={Durillo, Juan J and Nebro, Antonio J},
+  journal={Advances in Engineering Software},
+  volume={42},
+  number={10},
+  pages={760--771},
+  year={2011},
+  publisher={Elsevier},
+  doi={10.1016/j.advengsoft.2011.05.014}
+}
+
+@article{fortin:2012,
+  title={DEAP: Evolutionary algorithms made easy},
+  author={Fortin, F{\'e}lix-Antoine and De Rainville, Fran{\c{c}}ois-Michel and Gardner, Marc-Andr{\'e} Gardner and Parizeau, Marc and Gagn{\'e}, Christian},
+  journal={The Journal of Machine Learning Research},
+  volume={13},
+  number={1},
+  pages={2171--2175},
+  year={2012},
+  publisher={JMLR. org}
+}
+
+@article{cahon:2004,
+  title={Paradiseo: A framework for the reusable design of parallel and distributed metaheuristics},
+  author={Cahon, S{\'e}bastien and Melab, Nordine and Talbi, E-G},
+  journal={Journal of heuristics},
+  volume={10},
+  number={3},
+  pages={357--380},
+  year={2004},
+  publisher={Springer}
+}
+
+@article{fourer:2003,
+  title={AMPL. A modeling language for mathematical programming},
+  author={Fourer, Robert and Gay, David M and Kernighan, Brian W},
+  year={2003},
+  publisher={Thomson}
 }
