+++ paper.md
@@ -113,7 +113,7 @@ A large number of projects, open source as well as commercial, exist whose funct
 single-objective problems, some with (fine-grained) parallelization capabilities. On the other hand, projects like AMPL [@fourer:2003]
 or GAMS originate from the operational research community and offer modelling languages able to represent generic optimization 
 problems and to forward them, together with the jacobians and hessians needed, to compatible solvers. 
-A third type of projects, like NLOpt [@johnson:2014] or the Scipy optimize module offer a number of solvers without making much distinction between heuristic, derivative-free or local deterministic solvers. The project `pagmo`, has most of the capabilities of the above mentioned software integrated in the same ecosystem as it offers a large variety of parallelization modes and, above all, the possibility to code and wrap easily new (or third party) solvers, problems and parallelization strategies. A unique characteristic of `pagmo` to be here highlighted, is the presence of an island model implementation [@izzo:2012] that can flexibly distribute any solver, original, user implemented or third party wrapped on CPU clusters. 
+A third type of projects, like NLOpt [@johnson:2014] or the Scipy optimize module offer a number of solvers without making much distinction between heuristic, derivative-free or local deterministic solvers. The project `pagmo`, has most of the capabilities of the above mentioned softwares integrated in the same ecosystem as it offers a large variety of parallelization modes and, above all, the possibility to code and wrap easily new (or third party) solvers, problems and parallelization strategies. A unique characteristic of `pagmo` to be here highlighted, is the presence of an island model implementation [@izzo:2012] that can flexibly distribute any solver, original, user implemented or third party on CPU clusters. 
 
 # Code Design
 
