+++ paper.md
@@ -32,11 +32,15 @@ Given the vast amount and diversity of optimization problems, as well as of solu
 and considering the need to be able to exploit modern computational architectures, the development of 
 a tool able to help in such a pervasive task is not trivial.
 
-In this paper we introduce `pagmo`, a C++ scientific library for massively parallel optimization. `pagmo` is built around the idea of providing a unified interface to optimization algorithms and problems, and to make their deployment in massively parallel environments easy.
+In this paper we introduce `pagmo`, a C++ scientific library for massively parallel optimization. `pagmo` is built around the idea of providing
+a unified interface to optimization algorithms and problems, and to make their deployment in massively parallel environments easy.
 
-Efficient implementantions of bio-inspired and evolutionary algorithms are sided to state-of-the-art optimization algorithms (Simplex Methods, SQP methods, interior points methods, etc.) and can be used cuncurrently (also together with algorithms coded by the user) to build an optimization pipeline exploiting algorithmic cooperation via the asynchronous, generalized island model [@izzo:2012]
+Efficient implementantions of bio-inspired and evolutionary algorithms are sided to state-of-the-art optimization algorithms
+(Simplex Methods, SQP methods, interior points methods, etc.) and can be used concurrently (also together with algorithms coded by the user)
+to build an optimization pipeline exploiting algorithmic cooperation via the asynchronous, generalized island model [@izzo:2012].
 
-`pagmo` can be used to solve constrained, unconstrained, single objective, multiple objectives, continuous and integer optimization problems, stochastic and deterministic problems, as well as to perform research on novel algorithms and paradigms and easily compare them to state-of-the-art implementations of established ones.
+`pagmo` can be used to solve constrained, unconstrained, single objective, multiple objectives, continuous and integer optimization problems,
+stochastic and deterministic problems, as well as to perform research on novel algorithms and paradigms and easily compare them to state-of-the-art implementations of established ones.
 
 For users that are more comfortable with the Python language, the package `pygmo` following as closely as possible the `pagmo` API is also available.
 
@@ -56,8 +60,8 @@ where $\mathbf x \in \mathbb R^{n_{cx}} \times  \mathbb Z^{n_{ix}}$ is called *d
 total problem dimension is then indicated with $n_x = n_{cx} + n_{ix}$. $\mathbf{lb}, \mathbf{ub} \in
 \mathbb R^{n_{cx}} \times  \mathbb Z^{n_{ix}}$ are the *box-bounds*, $\mathbf f: \mathbb R^{n_{cx}} \times
 \mathbb Z^{n_{ix}} \rightarrow \mathbb R^{n_{obj}}$ define the *objectives*, $\mathbf c_e:  \mathbb R^{n_{cx}}
-\times  \mathbb Z^{n_{ix}} \rightarrow \mathbb R^{n_{ec}}$ are non linear *equality constraints*, and $\mathbf
-c_i:  \mathbb R^{n_{cx}} \times  \mathbb Z^{n_{ix}} \rightarrow \mathbb R^{n_{ic}}$ are non linear *inequality
+\times  \mathbb Z^{n_{ix}} \rightarrow \mathbb R^{n_{ec}}$ are nonlinear *equality constraints*, and $\mathbf
+c_i:  \mathbb R^{n_{cx}} \times  \mathbb Z^{n_{ix}} \rightarrow \mathbb R^{n_{ic}}$ are nonlinear *inequality
 constraints*. Note that the objectives and constraints also depend from an added value $s$ representing some
 stochastic variable. Both equality and inequality constraints are considered as satisfied whenever their definition
 is met within a tolerance $\mathbf c_{tol}$. 
[…]
