### 2020-08-25 #2338 @bluescarni
@eloisabentivegna @jangmys @dgoldri25 @darioizzo we have added a section about related projects/frameworks and fixed the typos. Please let us know if there is anything else.

@jangmys regarding your question about the GPU: what pagmo provides is an API to compute the fitness evaluation of a group of independent decision vectors in a (possibly) parallel fashion. Such API can then be used by optimisation algorithms capable of taking advantage of parallel fitness evaluation (in pagmo we have a handful of such algorithms, mostly genetic/evolutionary ones). Thus you are absolutely right that we don't directly address the specifics of GPU programming from within pagmo (such as data transfer, compilation model, etc.). This is left to the author of an optimisation problem, who is free to choose between CUDA/OpenCL/SYCL for the implementation of the batch evaluation API in his/her specific optimisation problem. We expanded a bit on this point in the latest revision of the paper, please let us know if this clarifies the matter.
