### 2020-08-27 #2338 @jangmys
@bluescarni @eloisabentivegna @dgoldri25 @darioizzo I think the additional paragraph addresses the existing state-of-the-art appropriately (and I therefore checked the corresponding box). I also agree with the "Concurrent fitness evaluation" section which  now clarifies that there is no "out-of-the-box" GPU support, but rather a clean interface that could allow users to build one on their own. From my side there are no further remarks and I recommend to accept the paper. I hope you'll be able to keep up the good work on this framework in the future ;-)
