### 2020-09-01 #2338 @eloisabentivegna
Thanks, @jangmys and @dgoldri25! It sounds like we are ready to accept the paper. @arfon, is there something special we need to do because of the double code repository?
