### 2020-09-01 #2338 @arfon
> @arfon, is there something special we need to do because of the double code repository?

We need the authors to make a single archive with e.g. Zenodo of all of the software associated with the submission. This will likely require a little extra work on the part of the authors compared with some of the automated methods for doing this with GitHub.
