### 2020-09-01 #2338 @bluescarni
@arfon would it be enough to make an archive of the latest releases of pagmo and pygmo (i.e., from the released tarballs on github)?
