### 2020-09-01 #2338 @arfon
> @arfon would it be enough to make an archive of the latest releases of pagmo and pygmo (i.e., from the released tarballs on github)?

We would like the archive to include any changes that have been made as a result of this review. Would that be the case if you used the latest releases?
