### 2020-09-03 #2338 @arfon
> @arfon where should the archive be uploaded?

We need an archive DOI from e.g. Zenodo or figshare.
