### 2020-09-05 #2338 @eloisabentivegna
@bluescarni, thanks for posting the archive DOI! The title should match the paper's, though. Could you change it to "A parallel global multiobjective framework for optimization: pagmo"?
