+++ paper.bib
@@ -172,7 +172,8 @@
   number={3},
   pages={357--380},
   year={2004},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1023/B:HEUR.0000026900.92269.ec}
 }
 
 @article{fourer:2003,
