### 2020-09-07 89ec718
Remove DOI prefix.
