+++ paper.bib
@@ -107,7 +107,7 @@
   pages={341--359},
   year={1997},
   publisher={Springer},
-  doi={https://doi.org/10.1023/A:1008202821328}
+  doi={10.1023/A:1008202821328}
 }
 
 @article{walt2011numpy,
