### 2020-09-08 #2338 @eloisabentivegna
@bluescarni, thanks for checking everything!

By a manual inspection, I noticed three further missing DOIs, and created a new pull request. @arfon, do you know why these three references might have not been picked up by the automated check?

Also @arfon, is it OK that the paper "Repository" link points to the paper repository, not the pagmo/pygmo ones? If not, is there a way to fix this?

Other than these points, the proofs look fine to me. @bluescarni, as soon as you confirm you are fine with the proofs too, I will complete the pre-publication stage.
