### 2020-09-08 #2338 @arfon
> By a manual inspection, I noticed three further missing DOIs, and created a new pull request. @arfon, do you know why these three references might have not been picked up by the automated check?

Whedon simply uses the Crossref API to check for missing DOIs, they presumably aren't being found there...

> Also @arfon, is it OK that the paper "Repository" link points to the paper repository, not the pagmo/pygmo ones? If not, is there a way to fix this?

It's not ideal but this is a consequence of having the paper in a separate repository. I would suggest directly linking to both variants of the software `pymgo` and `pagmo` in the body of the paper to make it easier for the reader.
