### 2020-09-10 #2338 @eloisabentivegna
@bluescarni, this will now pass to the editors in chief for the final steps. Thanks for submitting this great software to JOSS!

Just as importantly, thanks to @jangmys and @dgoldri25 for the thorough and insightful reviews!
