### 2020-09-10 #2338 @danielskatz
@bluescarni - a bunch of small changes in the paper:
* The last sentence in the summary has a grammar problem
* page 1, "is called decision" should be "is called the decision"
* page 2, "also depend from" should be "also depend on"
* "have no additional help to approach" isn't clear to me. Should this be "do not help the user to approach"?  Or something different?
* "referred to also as chromosome" should be "referred to also as a chromosome"
* "Regardless on" should be "Regardless of"
* Please refer to the user as "the user" or "they" rather than "he"
* "concurring to the same optimization" doesn't make sense to me, and I'm not able to suggest different language, as I don't understand what this should say
* darwinian should be Darwinian
* "parallelize at a fine grain the evaluation pipeline" should be "parallelize the evaluation pipeline at a fine grain"
* "it up to" should be "it is up to"
* "project pagmo, has most" should be "project pagmo has most"
* "softwares" should be "software packages"
* "the possibility to code and wrap easily" should be "the ability to code and easily wrap"
* "to be here highlighted, is" should be "to be highlighted here is"
* "they allow to minimise" should be "they allow [someone] to minimize" - I'm not sure if this should be the user or perhaps the developer or the programmer...
* "to set explicitly the" should be "to explicitly set the"
* "printed to screen" should be "printed to the screen"
* Should there be URLs for the Jakob et al. and Johnson references?  Similarly Fourer at al.  And Hagberg
* Also for Hagberg, please add a space before (LANL)
* Storn & Price ref, please make the journal title Title Case: "Journal of Global Optimization"
