+++ paper.md
@@ -42,7 +42,8 @@ to build an optimization pipeline exploiting algorithmic cooperation via the asy
 `pagmo` can be used to solve constrained, unconstrained, single objective, multiple objectives, continuous and integer optimization problems,
 stochastic and deterministic problems, as well as to perform research on novel algorithms and paradigms and easily compare them to state-of-the-art implementations of established ones.
 
-For users that are more comfortable with the Python language, the package `pygmo` following as closely as possible the `pagmo` API is also available.
+For users that are more comfortable with the Python language, the package `pygmo` provides a complete set of Python bindings for `pagmo` closely following
+the C++ API.
 
 # The optimization problem 
 In `pagmo` optimization problems are considered to be in the form:
@@ -55,19 +56,19 @@ $$
                      & \mathbf {c}_i(\mathbf x, s) \le 0
   \end{array}
 $$
-where $\mathbf x \in \mathbb R^{n_{cx}} \times  \mathbb Z^{n_{ix}}$ is called *decision vector* or
+where $\mathbf x \in \mathbb R^{n_{cx}} \times  \mathbb Z^{n_{ix}}$ is called a *decision vector* or
 *chromosome*, and is made of $n_{cx}$ real numbers and $n_{ix}$ integers (all represented as doubles). The
 total problem dimension is then indicated with $n_x = n_{cx} + n_{ix}$. $\mathbf{lb}, \mathbf{ub} \in
 \mathbb R^{n_{cx}} \times  \mathbb Z^{n_{ix}}$ are the *box-bounds*, $\mathbf f: \mathbb R^{n_{cx}} \times
 \mathbb Z^{n_{ix}} \rightarrow \mathbb R^{n_{obj}}$ define the *objectives*, $\mathbf c_e:  \mathbb R^{n_{cx}}
 \times  \mathbb Z^{n_{ix}} \rightarrow \mathbb R^{n_{ec}}$ are nonlinear *equality constraints*, and $\mathbf
 c_i:  \mathbb R^{n_{cx}} \times  \mathbb Z^{n_{ix}} \rightarrow \mathbb R^{n_{ic}}$ are nonlinear *inequality
-constraints*. Note that the objectives and constraints also depend from an added value $s$ representing some
+constraints*. Note that the objectives and constraints also depend on an added value $s$ representing some
 stochastic variable. Both equality and inequality constraints are considered as satisfied whenever their definition
 is met within a tolerance $\mathbf c_{tol}$. 
 
 Note that there is no special treatment of a possible linear part of the *objectives* or *constraints*, and as such solvers in `pagmo`
-have no additional help to approach linear programming tasks.
+cannot take advantage of the special structure of linear programming tasks.
 
 Given the generic form used to represent a problem, `pagmo` is suitable to solve a broad range of optimization problems,
 ranging from single and multiobjective problems to box-bounded and nonlinearly constrained problems to stochastic problems to continuous, integer and mixed integer problems.
@@ -77,13 +78,13 @@ The discussion on the relation between artificial evolution and mathematical opt
 interesting one [@smith:1978]. In `pagmo` optimization, 
 of all types, is regarded as a form of evolution. Solving an optimization problem
 is, in `pagmo`, described as *evolving* a *population*. Each *decision vector* is thus referred to also
-as *chromosome* and its *fitness* is defined as a vector containing *objectives*, *equality constraints* and *inequality
-constraints* in this order. Regardless on whether the user is using, as a solver, a sequential quadratic programming approach, 
-an interior point optimizer an evolutionary strategy or some meta-heuristic, in `pagmo` he will always have to call a method called
+as a *chromosome* and its *fitness* is defined as a vector containing *objectives*, *equality constraints* and *inequality
[…]
