+++ paper.bib
@@ -1,81 +1,81 @@
 @incollection{izzo:2012,
-  title={The generalized island model},
-  author={Izzo, Dario and Ruci{\'n}ski, Marek and Biscani, Francesco},
-  booktitle={Parallel Architectures and Bioinspired Algorithms},
-  pages={151--169},
-  year={2012},
-  publisher={Springer},
-  doi={10.1007/978-3-642-28789-3_7}
+  title     = {The generalized island model},
+  author    = {Izzo, Dario and Ruci{\'n}ski, Marek and Biscani, Francesco},
+  booktitle = {Parallel Architectures and Bioinspired Algorithms},
+  pages     = {151--169},
+  year      = {2012},
+  publisher = {Springer},
+  doi       = {10.1007/978-3-642-28789-3_7}
 }
 
 
 @misc{audi:2020,
-  author       = {Dario Izzo and
+  author  = {Dario Izzo and
                   Francesco Biscani},
-  title        = {AuDi: v1.8},
-  year         = 2020,
-  journal      = {Zenodo},
-  version      = {v1.8},
-  doi          = {10.5281/zenodo.3702783},
-  url          = {https://doi.org/10.5281/zenodo.3702783}
+  title   = {AuDi: v1.8},
+  year    = 2020,
+  journal = {Zenodo},
+  version = {v1.8},
+  doi     = {10.5281/zenodo.3702783},
+  url     = {https://doi.org/10.5281/zenodo.3702783}
 }
 
 @article{smith:1978,
-  title={Optimization theory in evolution},
-  author={Smith, J Maynard},
-  journal={Annual Review of Ecology and Systematics},
-  volume={9},
-  number={1},
-  pages={31--56},
-  year={1978},
-  publisher={Annual Reviews 4139 El Camino Way, PO Box 10139, Palo Alto, CA 94303-0139, USA},
-  doi={10.1146/annurev.es.09.110178.000335}
+  title     = {Optimization theory in evolution},
+  author    = {Smith, J Maynard},
+  journal   = {Annual Review of Ecology and Systematics},
+  volume    = {9},
+  number    = {1},
+  pages     = {31--56},
+  year      = {1978},
+  publisher = {Annual Reviews 4139 El Camino Way, PO Box 10139, Palo Alto, CA 94303-0139, USA},
+  doi       = {10.1146/annurev.es.09.110178.000335}
 }
 
 @article{rucinski:2010,
-  title={On the impact of the migration topology on the island model},
-  author={Ruci{\'n}ski, Marek and Izzo, Dario and Biscani, Francesco},
-  journal={Parallel Computing},
-  volume={36},
-  number={10-11},
-  pages={555--571},
-  year={2010},
-  doi={10.1016/j.parco.2010.04.002},
-  publisher={Elsevier}
+  title     = {On the impact of the migration topology on the island model},
+  author    = {Ruci{\'n}ski, Marek and Izzo, Dario and Biscani, Francesco},
+  journal   = {Parallel Computing},
+  volume    = {36},
+  number    = {10-11},
+  pages     = {555--571},
+  year      = {2010},
+  doi       = {10.1016/j.parco.2010.04.002},
+  publisher = {Elsevier}
 }
 
 @article{tanese:1989,
-  title={Distributed genetic algorithms for function optimization},
-  author={Tanese, Reiko},
-  year={1989},
-  publisher={University of Michigan}
+  title     = {Distributed genetic algorithms for function optimization},
+  author    = {Tanese, Reiko},
+  year      = {1989},
+  publisher = {University of Michigan}
 }
 
 @misc{pybind11,
-   author = {Wenzel Jakob and Jason Rhinelander and Dean Moldovan},
-   year = {2017},
-   note = {https://github.com/pybind/pybind11},
[…]
