+++ paper.bib
@@ -101,7 +101,7 @@
 @article{storn1997differential,
   title     = {Differential evolution--a simple and efficient heuristic for global optimization over continuous spaces},
   author    = {Storn, Rainer and Price, Kenneth},
-  journal   = {{Journal of global optimization}},
+  journal   = {{Journal of Global Optimization}},
   volume    = {11},
   number    = {4},
   pages     = {341--359},
