### 2020-09-13 #2338 @bluescarni
@eloisabentivegna @dgoldri25 @jangmys Thanks for the review and all the help!

@danielskatz Thanks for the review, I updated the paper with your corrections/suggestions. Regarding the missing links for the Jakob and Johnson references, I have used the bibtex entries suggested in the documentation of those software projects. Perhaps those entries don't play well with JOSS's bib style?
