### 2020-09-13 #2338 @danielskatz
Unfortunately, bibtex entries are often built for particular style, and don't work for other style.  I think, rather than putting URLs in a note field, you should put them in a url field (going from https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography)  Can you try this?
