+++ paper.bib
@@ -54,7 +54,7 @@
 @misc{pybind11,
   author = {Wenzel Jakob and Jason Rhinelander and Dean Moldovan},
   year   = {2017},
-  note   = {https://github.com/pybind/pybind11},
+  url    = {https://github.com/pybind/pybind11},
   title  = {pybind11 -- Seamless operability between C++11 and Python}
 }
 
@@ -62,7 +62,7 @@
   title  = {The NLopt nonlinear-optimization package},
   author = {Johnson, Steven G},
   year   = {2014},
-  note   = {http://github.com/stevengj/nlopt}
+  url    = {http://github.com/stevengj/nlopt}
 }
 
 @misc{ipyparallel,
