### 2020-09-13 #2338 @bluescarni
@danielskatz Cheers for the suggestion, it seems like now the URLs are showing up in the bibliography.
